% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer.dist.R
\docType{methods}
\name{buffer.dist,SpatialPointsDataFrame,SpatialPixelsDataFrame-method}
\alias{buffer.dist,SpatialPointsDataFrame,SpatialPixelsDataFrame-method}
\title{Derive buffer distances for a list of points}
\usage{

  \S4method{buffer.dist}{SpatialPointsDataFrame,SpatialPixelsDataFrame}(observations,
  predictionDomain, classes, width, parallel = TRUE, ...)
}
\arguments{
\item{observations}{SpatialPointsDataFrame.}

\item{predictionDomain}{SpatialPixelsDataFrame.}
}
\value{
object of class SpatialPixelsDataFrame
}
\description{
Derive buffer distances using the \code{raster::distance} function, so that these can be used as predictors for spatial prediction i.e. to account for spatial proximity to low, medium and high values.
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
