% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.vgmModel.R
\name{fit.vgmModel,formula,data.frame,SpatialPixelsDataFrame-method}
\alias{fit.vgmModel,formula,data.frame,SpatialPixelsDataFrame-method}
\title{Fit variogram using point data}
\usage{
\S4method{fit.vgmModel}{formula,data.frame,SpatialPixelsDataFrame}(
  formulaString.vgm,
  rmatrix,
  predictionDomain,
  cov.model = "exponential",
  dimensions = list("2D", "3D", "2D+T", "3D+T"),
  lambda = 0.5,
  psiR = NULL,
  subsample = nrow(rmatrix),
  ini.var,
  ini.range,
  fix.psiA = FALSE,
  fix.psiR = FALSE,
  ...
)
}
\arguments{
\item{formulaString.vgm}{formula.}

\item{rmatrix}{data.frame.}

\item{predictionDomain}{SpatialPixelsDataFrame.}
}
\value{

}
\description{
Fit variogram using point data
}
\note{
Extends variogram fitting functionality from the geoR package.
Can be used for 2D or 3D point data sets, with and without trend variables.
Models need to be in the form \code{zinc ~ dist} and only numeric variables are allowed.
}
\examples{
library(raster)
library(rgdal)
library(geoR)
demo(meuse, echo=FALSE)
vgm = fit.vgmModel(zinc~dist, as.data.frame(meuse), meuse.grid["dist"], lambda=1)
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
