% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialTiles.R
\docType{methods}
\name{getSpatialTiles,ANY-method}
\alias{getSpatialTiles,ANY-method}
\title{Estimate a tiling system}
\usage{
\S4method{getSpatialTiles}{ANY}(obj, block.x, block.y = block.x,
  overlap.percent = 0, limit.bbox = TRUE,
  return.SpatialPolygons = FALSE)
}
\arguments{
\item{obj}{ANY.}
}
\description{
Estimate a tiling system
}
