% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialTiles.R
\name{getSpatialTiles,Spatial-method}
\alias{getSpatialTiles,Spatial-method}
\title{Split a Spatial object into tiles}
\usage{
\S4method{tile}{SpatialPointsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPixelsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPolygonsDataFrame}(x, y, block.x, tmp.file = TRUE,
                                          program, show.output.on.console = FALSE, \dots)
\S4method{tile}{SpatialLinesDataFrame}(x, y, block.x, tmp.file = TRUE,
                                       program, show.output.on.console = FALSE, \dots)
\S4method{tile}{RasterLayer}(x, y, block.x, tmp.file = TRUE,
                             program, show.output.on.console = FALSE, \dots)
}
\arguments{
\item{obj}{Spatial.}
}
\value{

}
\description{
Split a Spatial object into tiles
}
\examples{
library(sp)
data(meuse.grid)
gridded(meuse.grid) <- ~x+y
tl <- getSpatialTiles(meuse.grid, block.x=1000)
image(meuse.grid)
lines(as(tl, "SpatialLines"))
## all at once:
pix.lst <- tile(meuse.grid, block.x=1000)
\dontrun{## lines:
  library(plotKML)
  data(eberg_contours)
  line.lst <- tile(eberg_contours, block.x=5000)
  spplot(line.lst[[1]][2])
  ## polygons:
  data(eberg_zones)
  ## this one requires ogr2ogr function:
  pol.lst <- tile(eberg_zones, block.x=5000)
  spplot(pol.lst[[1]][1])
  ## raster files via rgdal:
  library(rgdal)
  fn = system.file("pictures/SP27GTIF.TIF",
                   package = "rgdal")
  obj <- GDALinfo(fn)
  ras.lst <- getSpatialTiles(obj, block.x=1000)
  offset <- c(ras.lst$offset.y[1], ras.lst$offset.x[1])
  region.dim <- c(ras.lst$region.dim.y[1],
                  ras.lst$region.dim.x[1])
  ## read the first tile:
  SP27GTIF_T1 <- readGDAL(fn, offset=offset,
                          region.dim=region.dim)
  str(SP27GTIF_T1)
}
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
