% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{Make a tiling system from a bounding box}
\usage{
makeTiles(bb, block.x, block.y, overlap.percent, limit.bbox,
  columns = NULL, rows = NULL)
}
\arguments{
\item{rows}{}
}
\description{
Make a tiling system from a bounding box
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
