% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.spLearner.R
\name{model.data}
\alias{model.data}
\title{Overlay points and grids and prepare regression matrix}
\usage{
model.data(observations, formulaString, covariates, dimensions = c("2D",
  "3D", "2D+T", "3D+T"))
}
\arguments{
\item{dimensions}{}
}
\description{
Overlay points and grids and prepare regression matrix
}
