% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landgis.R
\name{search.landgis}
\alias{search.landgis}
\title{Search for available landgis layers}
\usage{
search.landgis(pattern,
  layersURL = "https://landgisapi.opengeohub.org/query/layers",
  update = FALSE)
}
\arguments{
\item{update}{}
}
\description{
Search for available landgis layers
}
