% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{class}
\name{SpatialComponents-class}
\alias{SpatialComponents-class}
\title{Spatial Principal Components class}
\value{

}
\description{
Spatial Principal Components class
}
\section{Slots}{

\describe{
\item{\code{predicted}}{SpatialPixelsDataFrame.}

\item{\code{pca}}{list.}
}}

