% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{class}
\name{SpatialMemberships-class}
\alias{SpatialMemberships-class}
\title{SpatialMemberships class}
\description{
SpatialMemberships class
}
\section{Slots}{

\describe{
\item{\code{predicted}}{SpatialPixelsDataFrame.}

\item{\code{model}}{list.}

\item{\code{mu}}{SpatialPixelsDataFrame.}

\item{\code{class.c}}{matrix.}

\item{\code{class.sd}}{matrix.}

\item{\code{confusion}}{ANY.}
}}

