% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAAA.R
\docType{class}
\name{spLearner-class}
\alias{spLearner-class}
\title{Spatial Prediction Learner class}
\description{
Spatial Prediction Learner class
}
\section{Slots}{

\describe{
\item{\code{spModel}}{ANY.}

\item{\code{vgmModel}}{list.}

\item{\code{covariates}}{SpatialPixelsDataFrame.}

\item{\code{spID}}{SpatialGridDataFrame.}
}}

