% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmultinom.R
\docType{methods}
\name{spmultinom,formula,SpatialPointsDataFrame,SpatialPixelsDataFrame-method}
\alias{spmultinom,formula,SpatialPointsDataFrame,SpatialPixelsDataFrame-method}
\title{Fits a multinomial logistic regression to spatial data}
\usage{

  \S4method{spmultinom}{formula,SpatialPointsDataFrame,SpatialPixelsDataFrame}(formulaString,
  observations, covariates, class.stats = TRUE, predict.probs = TRUE,
  ...)
}
\arguments{
\item{formulaString}{formula.}

\item{observations}{SpatialPointsDataFrame.}

\item{covariates}{SpatialPixelsDataFrame.}
}
\description{
Fits a multinomial logistic regression to spatial data
}
