% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\docType{methods}
\name{tile,RasterLayer-method}
\alias{tile,RasterLayer-method}
\alias{tile}
\alias{tile,SpatialLinesDataFrame-method}
\alias{tile,SpatialPixelsDataFrame-method}
\alias{tile,SpatialPointsDataFrame-method}
\alias{tile,SpatialPolygonsDataFrame-method}
\title{Tile spatial layers}
\usage{
\S4method{tile}{SpatialPointsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPixelsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPolygonsDataFrame}(x, y, block.x, tmp.file = TRUE,
                                          program, show.output.on.console = FALSE, \dots)
\S4method{tile}{SpatialLinesDataFrame}(x, y, block.x, tmp.file = TRUE,
                                       program, show.output.on.console = FALSE, \dots)
\S4method{tile}{RasterLayer}(x, y, block.x, tmp.file = TRUE,
                             program, show.output.on.console = FALSE, \dots)
}
\arguments{
\item{x}{RasterLayer.}

\item{y}{either points, pixels, polygons or lines.}

\item{block.x}{size of the block in x direction.}

\item{tmp.file}{temporary file name.}

\item{program}{optional location of the gdalwarp.}

\item{show.output.on.console}{shows progress.}

\item{...}{optional argument.}
}
\value{
Regular tiling system
}
\description{
Tile spatial layers
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
