% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tile.R
\name{tile,RasterLayer-method}
\alias{tile,RasterLayer-method}
\title{Tile spatial layers}
\usage{
\S4method{tile}{SpatialPointsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPixelsDataFrame}(x, y, block.x, \dots)
\S4method{tile}{SpatialPolygonsDataFrame}(x, y, block.x, tmp.file = TRUE,
                                          program, show.output.on.console = FALSE, \dots)
\S4method{tile}{SpatialLinesDataFrame}(x, y, block.x, tmp.file = TRUE,
                                       program, show.output.on.console = FALSE, \dots)
\S4method{tile}{RasterLayer}(x, y, block.x, tmp.file = TRUE,
                             program, show.output.on.console = FALSE, \dots)
}
\arguments{
\item{x}{RasterLayer.}
}
\value{

}
\description{
Tile spatial layers
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
