% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpatialTiles.R
\name{getSpatialTiles,ANY-method}
\alias{getSpatialTiles,ANY-method}
\title{Estimate a tiling system}
\usage{
\S4method{getSpatialTiles}{ANY}(
  obj,
  block.x,
  block.y = block.x,
  overlap.percent = 0,
  limit.bbox = TRUE,
  return.SpatialPolygons = FALSE
)
}
\arguments{
\item{obj}{ANY.}

\item{block.x}{size of the block in x dimension.}

\item{block.y}{size of the block in y dimension.}

\item{overlap.percent}{optional overlap percent between tiles.}

\item{limit.bbox}{optional bounding box.}

\item{return.SpatialPolygons}{logical specificies whether to return a data frame or Spatial Polygons.}
}
\value{
Tiling system for a spatial object
}
\description{
Estimate a tiling system
}
