% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTiles.R
\name{makeTiles}
\alias{makeTiles}
\title{Make a tiling system from a bounding box}
\usage{
makeTiles(
  bb,
  block.x,
  block.y,
  overlap.percent,
  limit.bbox,
  columns = NULL,
  rows = NULL
)
}
\arguments{
\item{bb}{Bounding Box}

\item{block.x}{Size of the block in X}

\item{block.y}{Size of the block in Y}

\item{overlap.percent}{Percent of overlap; default 0}

\item{limit.bbox}{Optional limiting bounding box}

\item{columns}{Optional number of columns}

\item{rows}{Optional number of rows}
}
\value{
A regular tiling system
}
\description{
Make a tiling system from a bounding box
}
\author{
\href{https://opengeohub.org/people/tom-hengl}{Tom Hengl}
}
