% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.spLearner.R
\name{predict.spLearner}
\alias{predict.spLearner}
\title{Predict spLearner}
\usage{
predict.spLearner(object, predictionLocations, model.error = TRUE, ...)
}
\arguments{
\item{...}{}
}
\value{
SpatialPixelsDataFrame object with predictions and model error.
}
\description{
Predict spLearner
}
