% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.spLearner.R
\name{print.spLearner}
\alias{print.spLearner}
\title{Print object of type 'spLearner'}
\usage{
\method{print}{spLearner}(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Print object of type 'spLearner'
}
