% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landgis.R
\name{search.landgis}
\alias{search.landgis}
\title{Search for available landgis layers}
\usage{
search.landgis(
  pattern,
  layersURL = "https://landgisapi.opengeohub.org/query/layers",
  update = FALSE
)
}
\arguments{
\item{pattern}{String pattern}

\item{layersURL}{Default URL with the list of layers}

\item{update}{Logical specify to update the layer list}
}
\value{
List of available landgis layers
}
\description{
Search for available landgis layers
}
